requirements.txt
Flask==2.2.2
numpy==1.23.5
Pillow==9.4.0
tensorflow==2.11.0
tensorflow_hub==0.12.0
tensorflow_intel==2.11.0
